% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{crossover_single_point}
\alias{crossover_single_point}
\title{Single-point crossover for vectors}
\usage{
crossover_single_point(parent1, parent2)
}
\arguments{
\item{parent1}{First parent vector.}

\item{parent2}{Second parent vector.}
}
\value{
A list of two offspring vectors.
}
\description{
Single-point crossover for vectors
}
