% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{CrossoverArithmetic}
\alias{CrossoverArithmetic}
\title{Arithmetic Crossover Class}
\description{
R6 class for arithmetic (real-valued) crossover.
}
\section{Super class}{
\code{geneOptimize::CrossoverOperator} -> \code{CrossoverArithmetic}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alpha}}{Mixing parameter.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossoverArithmetic-new}{\code{CrossoverArithmetic$new()}}
\item \href{#method-CrossoverArithmetic-mate}{\code{CrossoverArithmetic$mate()}}
\item \href{#method-CrossoverArithmetic-clone}{\code{CrossoverArithmetic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossoverArithmetic-new"></a>}}
\if{latex}{\out{\hypertarget{method-CrossoverArithmetic-new}{}}}
\subsection{Method \code{new()}}{
Initialize.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossoverArithmetic$new(alpha = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Mixing parameter (default 0.5).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossoverArithmetic-mate"></a>}}
\if{latex}{\out{\hypertarget{method-CrossoverArithmetic-mate}{}}}
\subsection{Method \code{mate()}}{
Mate two parents.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossoverArithmetic$mate(p1, p2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p1}}{Parent 1.}

\item{\code{p2}}{Parent 2.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossoverArithmetic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossoverArithmetic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossoverArithmetic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
