% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{CrossoverSinglePoint}
\alias{CrossoverSinglePoint}
\title{Single-Point Crossover Class}
\description{
R6 class for single-point crossover.
}
\section{Super class}{
\code{geneOptimize::CrossoverOperator} -> \code{CrossoverSinglePoint}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossoverSinglePoint-mate}{\code{CrossoverSinglePoint$mate()}}
\item \href{#method-CrossoverSinglePoint-clone}{\code{CrossoverSinglePoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossoverSinglePoint-mate"></a>}}
\if{latex}{\out{\hypertarget{method-CrossoverSinglePoint-mate}{}}}
\subsection{Method \code{mate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossoverSinglePoint$mate(p1, p2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossoverSinglePoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossoverSinglePoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossoverSinglePoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
