% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{CrossoverUniform}
\alias{CrossoverUniform}
\title{Uniform Crossover Class}
\description{
R6 class for uniform crossover.
}
\section{Super class}{
\code{\link[geneOptimize:CrossoverOperator]{geneOptimize::CrossoverOperator}} -> \code{CrossoverUniform}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossoverUniform-mate}{\code{CrossoverUniform$mate()}}
\item \href{#method-CrossoverUniform-clone}{\code{CrossoverUniform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossoverUniform-mate"></a>}}
\if{latex}{\out{\hypertarget{method-CrossoverUniform-mate}{}}}
\subsection{Method \code{mate()}}{
Mate two parents.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossoverUniform$mate(p1, p2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p1}}{Parent 1.}

\item{\code{p2}}{Parent 2.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossoverUniform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossoverUniform-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossoverUniform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
