% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{mutation_binary}
\alias{mutation_binary}
\title{Binary mutation (bit-flip)}
\usage{
mutation_binary(chromosome, mutation_rate, ...)
}
\arguments{
\item{chromosome}{Binary vector.}

\item{mutation_rate}{Probability of flipping each bit.}

\item{...}{Additional arguments.}
}
\value{
Mutated binary vector.
}
\description{
Binary mutation (bit-flip)
}
