% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{mutation_real}
\alias{mutation_real}
\title{Real-valued mutation (Gaussian)}
\usage{
mutation_real(chromosome, mutation_rate, lower, upper, ...)
}
\arguments{
\item{chromosome}{Numeric vector.}

\item{mutation_rate}{Probability of mutating each gene.}

\item{lower}{Lower bound.}

\item{upper}{Upper bound.}

\item{...}{Additional arguments.}
}
\value{
Mutated numeric vector.
}
\description{
Real-valued mutation (Gaussian)
}
