% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{MutationSimple}
\alias{MutationSimple}
\title{Simple Mutation Class}
\description{
R6 class for standard mutation.
}
\section{Super class}{
\code{\link[geneOptimize:MutationOperator]{geneOptimize::MutationOperator}} -> \code{MutationSimple}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MutationSimple-mutate}{\code{MutationSimple$mutate()}}
\item \href{#method-MutationSimple-clone}{\code{MutationSimple$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MutationSimple-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-MutationSimple-mutate}{}}}
\subsection{Method \code{mutate()}}{
Mutate genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutationSimple$mutate(genes, type, rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Gene vector.}

\item{\code{type}}{Type.}

\item{\code{rate}}{Rate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MutationSimple-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MutationSimple-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutationSimple$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
