% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{new_gene_opt_res}
\alias{new_gene_opt_res}
\title{Constructor for GA result object}
\usage{
new_gene_opt_res(best_chromosome, best_fitness, history, call)
}
\arguments{
\item{best_chromosome}{Best chromosome found.}

\item{best_fitness}{Best fitness value.}

\item{history}{Fitness history.}

\item{call}{Original function call.}
}
\value{
An object of class \code{gene_opt_res}.
}
\description{
Constructor for GA result object
}
