% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromosome.R
\name{RealChromosome}
\alias{RealChromosome}
\title{Real-Valued Chromosome Class}
\description{
R6 class representing a real-valued chromosome.
}
\section{Super class}{
\code{\link[geneOptimize:Chromosome]{geneOptimize::Chromosome}} -> \code{RealChromosome}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RealChromosome-new}{\code{RealChromosome$new()}}
\item \href{#method-RealChromosome-mutate}{\code{RealChromosome$mutate()}}
\item \href{#method-RealChromosome-clone}{\code{RealChromosome$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geneOptimize" data-topic="Chromosome" data-id="copy"><a href='../../geneOptimize/html/Chromosome.html#method-Chromosome-copy'><code>geneOptimize::Chromosome$copy()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RealChromosome-new"></a>}}
\if{latex}{\out{\hypertarget{method-RealChromosome-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RealChromosome$new(genes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RealChromosome-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-RealChromosome-mutate}{}}}
\subsection{Method \code{mutate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RealChromosome$mutate(mutation_rate, lower, upper, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RealChromosome-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RealChromosome-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RealChromosome$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
