% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ga.R
\name{run_ga}
\alias{run_ga}
\title{Functional interface for Genetic Algorithm}
\usage{
run_ga(
  fitness_fn,
  n_genes,
  pop_size = 50,
  generations = 100,
  selection_fn = SelectionTournament$new(),
  crossover_fn = CrossoverSinglePoint$new(),
  mutation_fn = MutationSimple$new(),
  crossover_rate = 0.8,
  mutation_rate = 0.01,
  type = "binary",
  elitism_count = 1,
  verbose = TRUE,
  parallel = FALSE,
  cores = 1,
  ...
)
}
\arguments{
\item{fitness_fn}{Fitness function.}

\item{n_genes}{Number of genes.}

\item{pop_size}{Population size.}

\item{generations}{Generations.}

\item{selection_fn}{Selection operator object (R6).}

\item{crossover_fn}{Crossover operator object (R6).}

\item{mutation_fn}{Mutation operator object (R6).}

\item{crossover_rate}{Crossover rate.}

\item{mutation_rate}{Mutation rate.}

\item{type}{Encoding type.}

\item{elitism_count}{Elitism count.}

\item{verbose}{Verbose output.}

\item{parallel}{Parallel execution.}

\item{cores}{Number of cores.}

\item{...}{Additional arguments.}
}
\value{
Result object of class \code{gene_opt_res}.
}
\description{
A high-level wrapper to run the Genetic Algorithm.
}
