% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{selection_roulette}
\alias{selection_roulette}
\title{Roulette wheel selection}
\usage{
selection_roulette(population, fitness_values, ...)
}
\arguments{
\item{population}{Population matrix.}

\item{fitness_values}{Vector of fitness values.}

\item{...}{Additional arguments.}
}
\value{
Selected individual vector.
}
\description{
Roulette wheel selection
}
