% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{selection_tournament}
\alias{selection_tournament}
\title{Tournament selection}
\usage{
selection_tournament(population, fitness_values, k = 3, ...)
}
\arguments{
\item{population}{Population matrix.}

\item{fitness_values}{Vector of fitness values.}

\item{k}{Tournament size (default 3).}

\item{...}{Additional arguments.}
}
\value{
Selected individual vector.
}
\description{
Tournament selection
}
