% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{SelectionOperator}
\alias{SelectionOperator}
\title{Selection Operator Base Class}
\description{
Base R6 class for selection.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SelectionOperator-select}{\code{SelectionOperator$select()}}
\item \href{#method-SelectionOperator-clone}{\code{SelectionOperator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SelectionOperator-select"></a>}}
\if{latex}{\out{\hypertarget{method-SelectionOperator-select}{}}}
\subsection{Method \code{select()}}{
Select individuals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SelectionOperator$select(pop, n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pop}}{Population matrix.}

\item{\code{n}}{Number to select.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SelectionOperator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SelectionOperator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SelectionOperator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
