% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{SelectionRoulette}
\alias{SelectionRoulette}
\title{Roulette Selection Class}
\description{
R6 class for roulette wheel selection.
}
\section{Super class}{
\code{\link[geneOptimize:SelectionOperator]{geneOptimize::SelectionOperator}} -> \code{SelectionRoulette}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SelectionRoulette-select}{\code{SelectionRoulette$select()}}
\item \href{#method-SelectionRoulette-clone}{\code{SelectionRoulette$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SelectionRoulette-select"></a>}}
\if{latex}{\out{\hypertarget{method-SelectionRoulette-select}{}}}
\subsection{Method \code{select()}}{
Select an individual.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SelectionRoulette$select(population, fitness_values, n = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population}}{Population matrix.}

\item{\code{fitness_values}}{Fitness values.}

\item{\code{n}}{Number.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SelectionRoulette-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SelectionRoulette-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SelectionRoulette$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
